ChallengeTypeSelector component
